requirements.txt
pywebview>=5.0
sympy>=1.13
numpy
scipy
scikit-learn
CoolProp
jedi
pypandoc_binary
pyinstaller
typing_extensions